first stab at fixing celery task to cron
